Continuing with issue #215. shift x and y are now float. There's still a bug with SER files that are not upscaled in the right way.

+++ src/core/siril.h
@@ -309,7 +309,7 @@ struct preprocessing_data {
 
 /* registration data, exists once for each image and each layer */
 struct registration_data {
-	int shiftx, shifty;	// we could have a subpixel precision, but is it needed? saved
+	float shiftx, shifty;	// we could have a subpixel precision, but is it needed? saved
 	float rot_centre_x, rot_centre_y;	// coordinates for the rotation centre, saved
 	float angle;		// angle for the rotation, saved
 	fitted_PSF *fwhm_data;	// used in PSF/FWHM registration, not saved
